REGISTER SUMMARY: Support multiple register summary files
Added support for a manually administered register file to
the register summary generator.
Added a manually administered register summary file for the
Revision 1.3, Part 8, section 2.2 Additions to Existing Registers.
This section of the standard lacks the context required to
successfully automatically parse these 6 register fields, so
a manually adminstered (static) corrective file was the easiest
option to implement.

Added a "check_all_registers" script to generate the revision 1.3
registers summary and document.