REGISTER SUMMARY: Rev 1.3 Corrections for Part 9 Flow Control
Part 9 defines a flow control bit in the Port n Control CSRs, but it uses
the incorrect offset for that CSR (0x08).  As a result these bits are added
to the wrong STD_REG register.

This update drops the automatically parsed, but incorrect bit definition and
adds manual management of the bit for register block IDs 0x0001, 0x0002, 0x0003
and 0x0009.