REGISTER SUMMARY: Enhance parsing of Rev 3.2 Part 6 registers
The Rev 3.2 specifications define four new register blocks,
which use the same LP-Serial registers as in 1.3 and 2.2, but
place them at new offsets.  It also defines new registers to add
in the new register blocsk.  Parsing this condensed part of the
standard results in "UNKNOWN" block IDs for all these registers.

The changes in this update determine which registers/bit fields
appear and which register blocks, and adds the registers/bit
fields to the register summaries accordingly.