REGISTER SUMMARY: Rev 2.2 Block 1,2,3,9 support
Register blocks 0x0001, 0x002, 0x0003, and 0x0009 are similar in
the rev 2.2 standard to the rev 1.3 standard, but the documentation
consolidates the register descriptions separate from the block layouts.
As a result, there's some additional code required that determines the
block ID of the registers and spits out correct block(s) for each register
and register field that they belong to.